# hospital-surgeons/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hospital Surgeons API",
    "version": "1.0.0",
    "description": "API documentation for Hospital Surgeons application",
    "contact": {
      "name": "API Support",
      "email": "admin@example.com"
    }
  },
  "servers": [
    {
      "url": "https://hospital-surgoen.onrender.com",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/hospitals/{id}/assignments": {
      "get": {
        "summary": "Get all assignments for a hospital",
        "tags": [
          "Hospitals"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Hospital ID"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "accepted",
                "declined",
                "completed",
                "cancelled"
              ]
            },
            "description": "Filter by assignment status"
          },
          {
            "in": "query",
            "name": "priority",
            "schema": {
              "type": "string",
              "enum": [
                "low",
                "medium",
                "high",
                "urgent",
                "emergency"
              ]
            },
            "description": "Filter by priority"
          }
        ],
        "responses": {
          "200": {
            "description": "Assignments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}